# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactKind": {
        "enum": [
          "source",
          "mesh_glb",
          "mesh_obj",
          "texture",
          "preview"
        ],
        "title": "ArtifactKind",
        "type": "string"
      },
...
      "Asset": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/AssetArtifact"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Used"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "generation_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Seconds"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_bytes": {
            "default": 0,
            "title": "Input Bytes",
            "type": "integer"
          },
          "input_filename": {
            "title": "Input Filename",
            "type": "string"
          },
          "input_key": {
            "title": "Input Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "$ref": "#/components/schemas/GenerationParams"
          },
          "progress_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Message"
          },
          "status": {
            "$ref": "#/components/schemas/AssetStatus",
            "default": "pending"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          },
          "write_amplification": {
            "$ref": "#/components/schemas/WriteAmplification"
          }
        },
        "required": [
          "id",
          "name",
          "input_key",
          "input_filename",
          "created_at",
          "updated_at"
        ],
        "title": "Asset",
        "type": "object"
      },
      "AssetArtifact": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "size_human": {
            "title": "Size Human",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "kind",
          "key",
          "size_bytes",
          "size_human",
          "content_type"
        ],
        "title": "AssetArtifact",
        "type": "object"
      },
      "AssetCreateRequest": {
        "description": "Point the generator at an already-uploaded source image (the /generate\nform uploads it via the existing presign flow, then submits its key).",
        "properties": {
          "engine": {
            "$ref": "#/components/schemas/GenerationEngine",
            "default": "triposr"
          },
          "input_key": {
            "title": "Input Key",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "remove_background": {
            "default": true,
            "title": "Remove Background",
            "type": "boolean"
          },
          "texture_resolution": {
            "default": 1024,
            "title": "Texture Resolution",
            "type": "integer"
          }
        },
        "required": [
          "input_key"
        ],
        "title": "AssetCreateRequest",
        "type": "object"
      },
...
      "AssetStatus": {
        "enum": [
          "pending",
          "running",
          "complete",
          "failed"
        ],
        "title": "AssetStatus",
        "type": "string"
      },
      "AssetUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "AssetUpdateRequest",
        "type": "object"
      },
...
      "GenerationEngine": {
        "enum": [
          "triposr",
          "hunyuan3d",
          "procedural"
        ],
        "title": "GenerationEngine",
        "type": "string"
      },
      "GenerationParams": {
        "properties": {
          "engine": {
            "$ref": "#/components/schemas/GenerationEngine",
            "default": "triposr"
          },
          "remove_background": {
            "default": true,
            "title": "Remove Background",
            "type": "boolean"
          },
          "texture_resolution": {
            "default": 1024,
            "title": "Texture Resolution",
            "type": "integer"
          }
        },
        "title": "GenerationParams",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegenerateRequest": {
        "description": "Optionally change the engine/params on a re-run; bumps the version.",
        "properties": {
          "engine": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationEngine"
              },
              {
                "type": "null"
              }
            ]
          },
          "remove_background": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remove Background"
          },
          "texture_resolution": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Texture Resolution"
          }
        },
        "title": "RegenerateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WriteAmplification": {
        "description": "The sample's headline lesson: one small input fans out into many larger\nB2 objects. `ratio` is output bytes / input bytes.",
        "properties": {
          "input_bytes": {
            "default": 0,
            "title": "Input Bytes",
            "type": "integer"
          },
          "input_bytes_human": {
            "default": "0 B",
            "title": "Input Bytes Human",
            "type": "string"
          },
          "object_count": {
            "default": 0,
            "title": "Object Count",
            "type": "integer"
          },
          "output_bytes": {
            "default": 0,
            "title": "Output Bytes",
            "type": "integer"
          },
          "output_bytes_human": {
            "default": "0 B",
            "title": "Output Bytes Human",
            "type": "string"
          },
          "ratio": {
            "default": 0.0,
            "title": "Ratio",
            "type": "number"
          }
        },
        "title": "WriteAmplification",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local API for Image to 3D Asset Library: reconstruct a 3D mesh, texture maps, and a preview render from a single source image with a local open-source engine (TripoSR), and store every artifact in Backblaze B2 as a versioned, content-addressed 3D asset library. This contract documents the local API, not a hosted public endpoint.",
    "title": "Image to 3D Asset Library API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/assets": {
      "get": {
        "operationId": "list_assets_endpoint_assets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  },
                  "title": "Response List Assets Endpoint Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Assets Endpoint",
        "tags": [
          "assets"
        ]
      },
      "post": {
        "operationId": "create_asset_endpoint_assets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Asset Endpoint",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/assets/{asset_id}": {
      "delete": {
        "operationId": "delete_asset_endpoint_assets__asset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Asset Endpoint",
        "tags": [
          "assets"
        ]
      },
      "get": {
        "operationId": "get_asset_endpoint_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset Endpoint",
        "tags": [
          "assets"
        ]
      },
      "patch": {
        "operationId": "update_asset_endpoint_assets__asset_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Asset Endpoint",
        "tags": [
          "assets"
        ]
      }
    },
    "/assets/{asset_id}/regenerate": {
      "post": {
        "operationId": "regenerate_asset_endpoint_assets__asset_id__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Asset Endpoint",
        "tags": [
          "assets"
        ]
      }
    },
...
  }
}